[Artworks] Move into artist namespace.
