+++ lib/components/artist/artworks/artwork.js
@@ -4,9 +4,9 @@ import Relay from 'react-relay';
 import React from 'react-native';
 const { Text, View, StyleSheet } = React;
 
-import ImageView from '../opaque_image_view';
-import SerifText from '../text/serif';
-import colors from '../../../data/colors';
+import ImageView from '../../opaque_image_view';
+import SerifText from '../../text/serif';
+import colors from '../../../../data/colors';
 
 export default class Artwork extends React.Component {
   render() {
