+++ lib/components/artist/artworks/grid.js
@@ -13,7 +13,7 @@ import React from 'react-native';
 const { Dimensions, View, ScrollView, StyleSheet } = React;
 
 import Artwork from './artwork';
-import { metaphysicsURL } from '../../relay/config';
+import { metaphysicsURL } from '../../../relay/config';
 
 
 /**
