+++ lib/components/artist/artworks/index.js
@@ -4,12 +4,12 @@ import Relay from 'react-relay';
 import React from 'react-native';
 const { Text, View, StyleSheet } = React;
 
-import Separator from '../separator'
-import SerifText from '../text/serif';
+import Separator from '../../separator'
+import SerifText from '../../text/serif';
 import ArtworksGrid from './grid';
 const { artworksQuery } = ArtworksGrid;
 
-import colors from '../../../data/colors';
+import colors from '../../../../data/colors';
 
 class Artworks extends React.Component {
   constructor(props) {
