+++ lib/containers/artist.js
@@ -8,7 +8,7 @@ const { ScrollView, View } = React;
 import Header from '../components/artist/header';
 import About from '../components/artist/about';
 import Shows from '../components/artist/shows';
-import Artworks from '../components/artworks';
+import Artworks from '../components/artist/artworks';
 
 import TabView from '../components/tab_view';
 import type TabSelectionEvent from '../components/tab_view';
